+++ js/app-initializer.js
@@ -986,11 +986,14 @@ function initStepInteractiveElements() {
                     }
                     break;
                 case 'practice':
-                    // Agregar videos en el primer, segundo y último paso del módulo
+                    // Agregar videos en todos los pasos del módulo
                     if (currentStep === 0) { // Introducción a la Práctica
                         insertPracticeVideos(resourcesContainer.id);
                         insertWebsiteLinks(resourcesContainer.id);
                     } else if (currentStep === 1) { // Desarrollo Paso a Paso
+                        insertPracticeVideos(resourcesContainer.id);
+                        insertWebsiteLinks(resourcesContainer.id);
+                    } else if (currentStep === 2) { // Resultado Esperado y Solución de Problemas
                         insertPracticeVideos(resourcesContainer.id);
                         insertWebsiteLinks(resourcesContainer.id);
                     } else if (currentStep === 3) { // Recursos para Continuar Aprendiendo